refactor: finalise withdrawal
